Extend Collision struct
The first part of a 2 part update. Here the Collision struct is extended to add an index variable. This variable could potentially be used in several ways. 

One of the best uses for this variable is to use it to store an id for 'sub-objects' hit within an object. i.e. If a raycast hits a model that has several sub-meshes the id of the sub-mesh within the parent model can be stored. 

Then from script, we can use:

getWord( %raycast, 8 )

...to easily retrieve the stored id of the mesh that was hit with the raycast.

Part 2 of this update is in source/scene/sceneContainer.cpp